refactor: make exporting util modules in vvisp-utils/utils
