+++ packages/vvisp-utils/src/filterPrivateKey.js
