+++ packages/vvisp-utils/src/privateKeyToAddress.js
@@ -1,5 +1,5 @@
 module.exports = function(privateKey) {
-  const filterPrivateKey = require('./utils/filterPrivateKey');
+  const filterPrivateKey = require('./filterPrivateKey');
   privateKey = filterPrivateKey(privateKey);
   const web3 = require('./getWeb3')();
