+++ packages/vvisp-utils/test/filterPrivateKey.test.js
@@ -2,7 +2,7 @@ const chai = require('chai');
 const expect = chai.expect;
 chai.should();
 
-const filterPrivateKey = require('../../src/utils/filterPrivateKey');
+const { filterPrivateKey } = require('../src');
 
 const SAMPLE_PRIVATE_KEY =
   '8bb0722ff8cb8161da257dc2d3712a17db1753d1de2d8b6b27b0e4636d9899f6';
